Fader Canvas fix
When viewed on 16x10 resolution, the fader box was showing a gap at the bottom. It was set to 800x450 resolution but the screen scaler part of the HUD canvas was 800x600 with only horizontal scaling. As a result, it scaled wide but not tall. I changed the resolution on the fader rectangle to match the HUD canvas. Seems to work well at all resolutions.