Fixing the Bone viewer in Match 3
Problem 1: wanted to remove the weird rodent skull mesh that kept popping up when you right clicked. I found it was embedded in a couple of different objects, Bone under the BoneViewer Scene, and attached to ViewBone on the Match3 scene.

Problem 2: when you right clicked on a second object to view, it was just overlapping, not actually destroying the first one. It was giving an error, saying it couldn't destroy a transform of a game object.  so on line 82 I added .gameObject to Destroy(bone.transform.GetChild(i).gameObject); and that seems to work without giving us errors. But I was afraid of it, so I added a note there too :)